% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_combine.R
\name{clean_and_combine_raw_data}
\alias{clean_and_combine_raw_data}
\title{Clean and combine raw data}
\usage{
clean_and_combine_raw_data(save_in_app = T, re_clean_all = F)
}
\arguments{
\item{save_in_app}{Should you save a new version of the cleaned data into the Shiny app folder?}

\item{re_clean_all}{Should all raw data be put through the cleaning process or just the data that has not been cleaned thus far}
}
\description{
The folder for the raw data is defined by the get paths function. This function loads that data
then cleans it (reformats columns and extracts relevent data)
}
