% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update jobs database}
\usage{
update_database(civil_service_user = T)
}
\arguments{
\item{civil_service_user}{Are yoy a Civil Service User - if you are - you will need your username and password
in the working directory as an R file called user_name_and_password.R with the structure:
username = [CSJ username]
password = [CSJ password]}
}
\description{
Performs the Civil Service Jobs scrape - cleans the new data
}
